requirements.txt
feedparser==6.0.11
PyYAML==6.0.2
python-dateutil==2.9.0.post0
Jinja2==3.1.4
requests==2.32.3
